refactored test cases for BitVectors
